Refactor ome-startd.sh and ome.service for improved configuration and clarity

+++ ome/origin/ome-startd.sh
@@ -1,16 +1,31 @@
 #!/bin/bash
 
+# --- Configuration ---
+OME_DOCKER_HOME="/opt/ovenmediaengine"
+
+# --- Run OME Docker container ---
+echo "Checking environment variables...:"
+if [ -z "$OME_HOST_IP" ]; then
+  echo "❌ Missing required environment variables. Please set OME_HOST_IP."
+  exit 1
+fi
+
+# make sure that is not running
+echo "🧹Stopping and removing any existing OME container..."
 docker stop ome
 docker rm ome
+
+
+# --- Run OME Docker container ---
+echo "🚀 Launching OvenMediaEngine..."
+
+# RUN
+# docker run -d --name ome --restart unless-stopped \
 docker run -d --name ome \
   -e OME_HOST_IP="$OME_HOST_IP" \
-  -e PUBLIC_IPV4="$PUBLIC_IPV4" \
-  -e API_ACCESS_TOKEN="$API_ACCESS_TOKEN" \
-  -e OVT_WORKER_COUNT="$OVT_WORKER_COUNT" \
-  -e WEBRTC_SIGNALLING_WORKER_COUNT="$WEBRTC_SIGNALLING_WORKER_COUNT" \
-  -v "$OME_DOCKER_HOST/conf":/opt/ovenmediaengine/bin/origin_conf \
-  -v "$OME_DOCKER_HOST/logs":/var/log/ovenmediaengine \
-  -v "$OME_DOCKER_HOST/rec":/mnt/record \
+  -v "$OME_DOCKER_HOME/conf":/opt/ovenmediaengine/bin/origin_conf \
+  -v "$OME_DOCKER_HOME/logs":/var/log/ovenmediaengine \
+  -v "$OME_DOCKER_HOME/rec":/mnt/record \
   -p 1935:1935 \
   -p 9999:9999/udp \
   -p 9000:9000 \
@@ -22,4 +37,5 @@ docker run -d --name ome \
   -p 20081:20081 \
   -p 10000-10002:10000-10002/udp \
   airensoft/ovenmediaengine:latest
-echo "✅ OME is up and running!"
+
+echo "✅ OME  is up and running!"
